requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic>=2.6
pydantic-settings>=2.2
SQLAlchemy[asyncio]>=2.0
asyncpg>=0.29
python-dotenv>=1.0
alembic>=1.13.2
psycopg[binary]>=3.1
email-validator>=2.0
numpy>=1.26
PyYAML>=6.0
sentence-transformers>=2.7